backup/flask_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-cors==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.3
pytest==7.4.0
pytest-cov==4.1.0
